// backend/src/rps/types/gateway-events.type.ts
export enum RpsWsMessage {
    USER_CONNECTED = 'user-connected.message',
    LOAD_ARMOR_BUILDER = 'load-armor-builder.message',
    SAVE_BUILT_ARMOUR = 'save-built-armour.message',
    LOAD_LOCATION = 'load-location.message',
}

export enum RpsWsResponse {
    USER_CONNECTED = 'user-connected.response',
    LOAD_ARMOR_BUILDER = 'load-armor-builder.response',
    SAVE_BUILT_ARMOUR = 'save-built-armour.response',
    LOAD_LOCATION = 'load-location.response',
}